Extracted importance output parametrs
