+++ private/local_run-importance-ch22.sh
@@ -5,5 +5,8 @@ PATH=${PATH}:${PWD}/bin
 export VS_ECHO_CMDLINE=YES
 
 variant-spark --spark --master 'local[*]' -- \
- importance -if ${PWD}/data/chr22_1000.vcf -ff ${PWD}/data/chr22-labels.csv -fc 22_16050408 -v -rn 500 -rbs 20 -ro -sr 13 "$@"
+ importance -if ${PWD}/data/chr22_1000.vcf -ff ${PWD}/data/chr22-labels.csv -fc 22_16050408 -v -rn 500 -rbs 100 -sr 13 \
+ -of ${PWD}/tmp/importance-chr22.csv \
+ -ovn raw -ivb \
+ "$@"
